+++ docs/en/connector-v2/sink/SelectDB-Cloud.md
@@ -7,27 +7,33 @@
 Used to send data to SelectDB Cloud. Both support streaming and batch mode.
 The internal implementation of SelectDB Cloud sink connector upload after batch caching and commit the CopyInto sql to load data into the table.
 
+:::tip
+
+Version Supported
+
+* supported  `SelectDB Cloud version is >= 2.2.x`
+
+:::
+
 ## Key features
 
 - [x] [exactly-once](../../concept/connector-v2-features.md)
-
-By default, we use 2PC commit to ensure `exactly-once`
+- [x] [cdc](../../concept/connector-v2-features.md)
 
 ## Options
 
-|        name         |  type  | required |  default value  |
-|---------------------|--------|----------|-----------------|
-| load-url            | string | yes      | -               |
-| jdbc-url            | string | yes      | -               |
-| cluster-name        | string | yes      | -               |
-| username            | string | yes      | -               |
-| password            | string | yes      | -               |
-| table.identifier    | string | yes      | -               |
-| selectdb.config     | map    | yes      | -               |
-| sink.buffer-size    | int    | no       | 1024*1024 (1MB) |
-| sink.buffer-count   | int    | no       | 3               |
-| sink.max-retries    | int    | no       | 1               |
-| sink.check-interval | int    | no       | 10000           |
+|       name        |  type  | required |     default value      |
+|-------------------|--------|----------|------------------------|
+| load-url          | string | yes      | -                      |
+| jdbc-url          | string | yes      | -                      |
+| cluster-name      | string | yes      | -                      |
+| username          | string | yes      | -                      |
+| password          | string | yes      | -                      |
+| table.identifier  | string | yes      | -                      |
+| selectdb.config   | map    | yes      | -                      |
+| sink.buffer-size  | int    | no       | 10 * 1024 * 1024 (1MB) |
+| sink.buffer-count | int    | no       | 10000                  |
+| sink.max-retries  | int    | no       | 3                      |
 
 ### load-url [string]
 
@@ -56,33 +62,36 @@ The name of `SelectDB Cloud` table, the format is `database.table`
 ### sink.properties [string]
 
 Write property configuration
+
 CSV Write：
+
+```
 selectdb.config {
-file.type='csv'
-file.column_separator=','
-file.line_delimiter='\n'
+    file.type="csv"
+    file.column_separator=","
+    file.line_delimiter="\n"
 }
+```
+
 JSON Write:
+
+```
 selectdb.config {
-file.type="json"
-file.strip_outer_array="false"
+    file.type="json"
 }
+```
 
 ### sink.buffer-size [string]
 
-Write data cache buffer size, unit byte. The default is 1 MB, and it is not recommended to modify it.
+The maximum capacity of the cache, in bytes, that is flushed to the object storage. The default is 10MB. it is not recommended to modify it.
 
 ### sink.buffer-count [string]
 
-The number of write data cache buffers, the default is 3, it is not recommended to modify.
+Maximum number of entries flushed to the object store. The default value is 10000. it is not recommended to modify.
 
 ### sink.max-retries [string]
 
-The maximum number of retries in the Commit phase, the default is 1.
-
-### sink.check-interval [string]
-
-Periodic interval for writing files, in milliseconds, default 10 seconds.
+The maximum number of retries in the Commit phase, the default is 3.
 
 ## Example
 
@@ -99,7 +108,6 @@ sink {
     password="******"
     selectdb.config {
         file.type="json"
-        file.strip_outer_array="false"
     }
   }
 }
@@ -117,9 +125,9 @@ sink {
     username="admin"
     password="******"
     selectdb.config {
-        file.type='csv' 
-        file.column_separator=',' 
-        file.line_delimiter='\n' 
+        file.type="csv"
+        file.column_separator="," 
+        file.line_delimiter="\n" 
     }
   }
 }
