--- seatunnel-connectors-v2/connector-selectdb-cloud/src/main/java/org/apache/seatunnel/connector/selectdb/serialize/SelectDBDelimiterParser.java
@@ -1,81 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.seatunnel.connector.selectdb.serialize;
-
-import org.apache.seatunnel.common.exception.CommonErrorCode;
-import org.apache.seatunnel.connector.selectdb.exception.SelectDBConnectorException;
-
-import com.google.common.base.Strings;
-
-import java.io.StringWriter;
-
-public class SelectDBDelimiterParser {
-    private static final int SHIFT = 4;
-
-    private static final String HEX_STRING = "0123456789ABCDEF";
-
-    public static String parse(String sp, String dSp) throws RuntimeException {
-        if (Strings.isNullOrEmpty(sp)) {
-            return dSp;
-        }
-        if (!sp.toUpperCase().startsWith("\\X")) {
-            return sp;
-        }
-        String hexStr = sp.substring(2);
-        // check hex str
-        if (hexStr.isEmpty()) {
-            throw new SelectDBConnectorException(
-                    CommonErrorCode.ILLEGAL_ARGUMENT,
-                    "Failed to parse delimiter: `Hex str is empty`");
-        }
-        if (hexStr.length() % 2 != 0) {
-            throw new SelectDBConnectorException(
-                    CommonErrorCode.ILLEGAL_ARGUMENT,
-                    "Failed to parse delimiter: `Hex str is empty`");
-        }
-        for (char hexChar : hexStr.toUpperCase().toCharArray()) {
-            if (HEX_STRING.indexOf(hexChar) == -1) {
-                throw new SelectDBConnectorException(
-                        CommonErrorCode.ILLEGAL_ARGUMENT,
-                        "Failed to parse delimiter: `Hex str is empty`");
-            }
-        }
-        // transform to separator
-        StringWriter writer = new StringWriter();
-        for (byte b : hexStrToBytes(hexStr)) {
-            writer.append((char) b);
-        }
-        return writer.toString();
-    }
-
-    private static byte[] hexStrToBytes(String hexStr) {
-        String upperHexStr = hexStr.toUpperCase();
-        int length = upperHexStr.length() / 2;
-        char[] hexChars = upperHexStr.toCharArray();
-        byte[] bytes = new byte[length];
-        for (int i = 0; i < length; i++) {
-            int pos = i * 2;
-            bytes[i] = (byte) (charToByte(hexChars[pos]) << SHIFT | charToByte(hexChars[pos + 1]));
-        }
-        return bytes;
-    }
-
-    private static byte charToByte(char c) {
-        return (byte) HEX_STRING.indexOf(c);
-    }
-}
